Standardizing the replacement for String.IsNullOrWhitespace when building for .NET 3.5.

+++ Sources/Extras/Accord.Video.FFMPEG.GPL/VideoFileWriter.cpp
@@ -565,7 +565,7 @@ namespace Accord {
                     }*/
                 }
 
-                void send_audio_frame(OutputStream* ost, uint8_t* current, size_t length)
+                void send_audio_frame(OutputStream* ost, uint8_t* current, int length)
                 {
                     AVCodecContext* c = ost->enc;
